# apps/server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Test System IO API
  version: 0.14.0
  description: |
    HTTP API for uploading and viewing Playwright, Cypress, and Detox test
    reports. This spec is hand-authored and is the source of truth for the
    contract.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local dev
  - url: https://tsio.example.com
    description: Production
...
security:
  - apiKey: []
  - bearer: []
  - sessionCookie: []
paths:
...
  /api/v1/auth/github:
    get:
      tags: [auth]
      summary: Begin GitHub OAuth — redirects the browser to the authorize URL
      security: []
      responses:
        '302':
          description: Redirect to GitHub OAuth
        '503':
          $ref: '#/components/responses/OAuthDisabled'
...
  /api/v1/auth/github/callback:
    get:
      tags: [auth]
      summary: GitHub OAuth callback; sets session cookie on success
      security: []
      parameters:
        - in: query
          name: code
          schema: { type: string }
          required: true
        - in: query
          name: state
          schema: { type: string }
          required: true
      responses:
        '302':
          description: Redirect to the web client
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the session refresh token
      security: []
      responses:
        '204':
          description: Access cookie refreshed
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/oidc-policies:
    post:
      tags: [auth]
      summary: Install a GitHub OIDC policy (admin-key gated)
      security: []
      parameters:
        - in: header
          name: X-Admin-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repository_pattern, role]
              properties:
                repository_pattern:
                  type: string
                  description: '"owner/*", "owner/repo", or "*".'
                role:
                  type: string
                  description: contributor|uploader|viewer|editor|admin
                description:
                  type: string
      responses:
        '201':
          description: Policy installed
          content:
            application/json:
              schema:
                type: object
                required: [id, role]
                properties:
                  id: { type: string, format: uuid }
                  role: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/reports:
    get:
      tags: [reports]
      summary: Paginated report list
      security: []
      parameters:
        - in: query
          name: group
          schema: { type: string }
          description: Filter by report group slug
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Offset-paginated list of reports
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportListResponse' }

    post:
      tags: [reports]
      summary: Upload a test report bundle
      parameters:
        - in: header
          name: X-Report-Idempotency-Key
          schema: { type: string, minLength: 16, maxLength: 128 }
          required: false
        - in: header
          name: X-Report-Source
          schema: { type: string }
          required: false
        - in: header
          name: X-Report-Framework
          schema: { type: string, enum: [playwright, cypress, detox, maestro] }
          required: false
        - in: header
          name: X-Report-Repository
          schema: { type: string }
          required: false
          description: "GitHub 'owner/repo' slug."
        - in: header
          name: X-Report-Branch
          schema: { type: string }
          required: false
        - in: header
          name: X-Report-Commit
          schema: { type: string }
          required: false
        - in: header
          name: X-Report-GH-Run-ID
          schema: { type: string }
          required: false
        - in: header
          name: X-Report-GH-Run-Number
          schema: { type: string }
          required: false
        - in: header
          name: X-Report-GH-Run-Attempt
          schema: { type: string }
          required: false
          description: "Default '1' if omitted."
        - in: header
          name: X-Report-GH-PR-Number
          schema: { type: integer }
          required: false
        - in: header
          name: X-Report-GH-Job-ID
          schema: { type: string }
          required: false
        - in: header
          name: X-Report-GH-Job-Name
          schema: { type: string }
          required: false
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [bundle]
              properties:
                bundle:
                  type: string
                  format: binary
                metadata:
                  type: string
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '200':
          description: Idempotent replay — existing report returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
...
  /api/v1/reports/individual:
    get:
      tags: [reports]
      summary: Flat per-report list with test stats
      security: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Paginated individual reports
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IndividualReportListResponse' }
  /api/v1/reports/consolidated:
    get:
      tags: [reports]
      summary: Consolidated cross-run results for a filter (repo + branch + commit + name)
      security: []
      parameters:
        - in: query
          name: repository
          schema: { type: string }
          required: true
        - in: query
          name: branch
          schema: { type: string }
          required: true
        - in: query
          name: commit
          schema: { type: string }
          required: true
        - in: query
          name: name
          schema: { type: string }
          required: true
        - in: query
          name: run_attempt
          schema: { type: integer, minimum: 1 }
        - in: query
          name: gid
          schema: { type: string }
      responses:
        '200':
          description: Consolidated spec results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsolidatedResultsResponse' }
  /api/v1/reports/{id}:
    parameters:
      - in: path
        name: id
        schema: { type: string, format: uuid }
        required: true
    get:
      tags: [reports]
      security: []
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [admin]
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/reports/{id}/suites:
    parameters:
      - in: path
        name: id
        schema: { type: string, format: uuid }
        required: true
    get:
      tags: [reports]
      security: []
      responses:
        '200':
          description: Suites for the report
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Suite' }
  /api/v1/reports/{id}/suites/{suiteID}/specs:
    parameters:
      - in: path
        name: id
        schema: { type: string, format: uuid }
        required: true
      - in: path
        name: suiteID
        schema: { type: string, format: uuid }
        required: true
    get:
      tags: [reports]
      security: []
      responses:
        '200':
          description: Flat spec list under the suite
          content:
            application/json:
              schema:
                type: object
                required: [specs]
                properties:
                  specs:
                    type: array
                    items: { $ref: '#/components/schemas/TestSpec' }
  /api/v1/reports/{id}/cases:
    parameters:
      - in: path
        name: id
        schema: { type: string, format: uuid }
        required: true
      - in: query
        name: status
        schema:
          type: string
          enum: [passed, failed, skipped, flaky, timedOut, interrupted]
      - in: query
        name: suite_id
        schema: { type: string, format: uuid }
    get:
      tags: [reports]
      security: []
      responses:
        '200':
          description: Test cases for the report
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TestCase' }
  /api/v1/reports/{id}/search:
    parameters:
      - in: path
        name: id
        schema: { type: string, format: uuid }
        required: true
      - in: query
        name: q
        schema: { type: string, minLength: 1 }
        required: true
        description: Search query. Must be at least `search_min_length` chars from /api/v1/config.
      - in: query
        name: limit
        schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
    get:
      tags: [reports]
      security: []
      responses:
        '200':
          description: Suite-grouped matches
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/reports/{id}/json:
    parameters:
      - in: path
        name: id
        schema: { type: string, format: uuid }
        required: true
    get:
      tags: [reports]
      security: []
      summary: Download the raw Playwright JSON (redirected to a presigned S3 URL)
      responses:
        '302': { description: Redirect to presigned S3 URL }

  # ─── Test shard orchestration ──────────────────────────────────────────────
  # Live progress events are published onto the existing internal/events Hub
  # and delivered over /api/v1/ws. Subscribers send a frame
  #   { type: "subscribe.orchestration", identity: <CompositeIdentity> }
  # to receive only events for that run. Event types:
  #   - orchestration.run.started
  #   - orchestration.unit.leased
  #   - orchestration.unit.completed
  #   - orchestration.lease.expired
  #   - orchestration.run.completed
  #   - orchestration.run.timed_out
...
  /api/v1/artifacts/{id}:
    parameters:
      - in: path
        name: id
        schema: { type: string, format: uuid }
        required: true
    get:
      tags: [artifacts]
      responses:
        '302': { description: Redirect to presigned S3 URL }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/tests/history:
    get:
      tags: [tests]
      summary: Outcome series for one test across commits, newest first
      description: |
        Records whether the test failed in earlier runs. Returns one entry per report group that ran
        the test, plus a precomputed summary (flip count, failure rate, last
        passing commit, and the oldest commit in the current failing streak) so
        every caller derives those the same way. Summary counts and streaks cover
        the full requested time window even when entries are limited. Historical
        failures alone do not establish that a PR has no effect on a test.
      security: []
      parameters:
        - in: query
          name: test_id
          description: The test's stable_key — its MM-T id (e.g. MM-T4783_1) or, for repositories without ids, its full title
          schema: { type: string }
          required: true
        - in: query
          name: repo
          description: Full "owner/repo" slug or just the trailing repo segment
          schema: { type: string }
          required: true
        - in: query
          name: branch
          description: Exact branch. master excludes pull request groups, including PRs whose head branch is named master.
          schema: { type: string }
        - in: query
          name: baseline
          description: Require a full owner/repo slug, master branch, no pull request, and completed groups with every declared report uploaded.
          schema: { type: boolean, default: false }
        - in: query
          name: name
          description: Exact report group name, for keeping different suites and editions separate.
          schema: { type: string }
        - in: query
          name: before
          description: Include only groups created at or before this timestamp.
          schema: { type: string, format: date-time }
        - in: query
          name: framework
          schema: { type: string, enum: [playwright, cypress, detox, maestro] }
        - in: query
          name: run_group
          schema: { type: string }
        - in: query
          name: window
          description: Lookback window from now (30d, 24h, 90m). Defaults to 30d; values above 180d are rejected.
          schema: { type: string, pattern: '^[0-9]+[dhm]$', default: '30d' }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        '200':
          description: Test history series
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestHistoryResponse' }
  /api/v1/tests/evidence:
    get:
      tags: [tests]
      summary: One run's failures with error, stack and screenshots, grouped by cause
      description: |
        Everything stored about the failures of one run: error message, stack,
        screenshots, retry count, and the `stable_key` each failure is looked
        up by in `/tests/history`. Failures are grouped by normalized error
        text so 300 identical "element not visible" failures read as one
        cluster with 300 members.

        Identify the run with `group_id` or with the same composite identity
        the upload used: repository + commit_sha + gh_run_id + name
        (+ gh_run_attempt, default "1").

        No judgement is made here. Whether a failure is flaky or a regression
        is for the consumer to decide.
      security: []
      parameters:
        - in: query
          name: group_id
          schema: { type: string, format: uuid }
        - in: query
          name: repository
          schema: { type: string }
        - in: query
          name: commit_sha
          schema: { type: string }
        - in: query
          name: gh_run_id
          schema: { type: string }
        - in: query
          name: name
          schema: { type: string }
        - in: query
          name: gh_run_attempt
          schema: { type: string, default: '1' }
      responses:
        '200':
          description: Evidence pack
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestEvidence' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/triage/attribution:
    get:
      tags:
      - triage
      summary: Preview conservative attribution for every observed failure in one exact run
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageAssessment'
        '400':
          description: Invalid request
        '401':
          description: A dedicated triage key or allowlisted workflow is required
        '403':
          description: The workflow is not authorized for this repository
        '404':
          description: Not found
        '409':
          description: Lease, state, provenance or policy conflict
        '503':
          description: Optional external integration disabled or unavailable
      parameters:
      - in: query
        name: repository
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$
      - in: query
        name: commit_sha
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 64
          pattern: ^(?:[a-fA-F0-9]{40}|[a-fA-F0-9]{64})$
      - in: query
        name: gh_run_id
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 20
      - in: query
        name: gh_run_attempt
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 20
      - in: query
        name: name
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
  /api/v1/triage/assessments:
    post:
      tags:
      - triage
      summary: Append a durable shadow assessment; it cannot authorize a status override
      security:
      - triageKey: []
      - bearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageAssessment'
        '400':
          description: Invalid request
        '401':
          description: A dedicated triage key or allowlisted workflow is required
        '403':
          description: The workflow is not authorized for this repository
        '404':
          description: Not found
        '409':
          description: Lease, state, provenance or policy conflict
        '503':
          description: Optional external integration disabled or unavailable
        '201':
          description: Recorded before response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageAssessment'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriageRunSelector'
  /api/v1/triage/verdicts/{id}:
    get:
      tags:
      - triage
      summary: Read the immutable assessment recorded at decision time
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageAssessment'
        '400':
          description: Invalid request
        '401':
          description: A dedicated triage key or allowlisted workflow is required
        '403':
          description: The workflow is not authorized for this repository
        '404':
          description: Not found
        '409':
          description: Lease, state, provenance or policy conflict
        '503':
          description: Optional external integration disabled or unavailable
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
...
components:
  securitySchemes:
    triageKey:
      type: apiKey
      in: header
      name: X-Triage-Key
      description: Dedicated triage authority; ordinary upload keys and sessions are refused.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API-key auth for CI upload clients.
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: GitHub Actions OIDC ID token for CI workload auth.
    sessionCookie:
      type: apiKey
      in: cookie
      name: tsio_session
      description: Opaque server-side session for human users.
  schemas:
...
    ReportDetail:
      allOf:
        - $ref: '#/components/schemas/ReportSummary'
        - type: object
          required: [updated_at, reports]
          properties:
            updated_at: { type: string, format: date-time }
            reports:
              type: array
              items: { $ref: '#/components/schemas/ReportEntry' }
            error_message: { type: string }
...
    ReportListResponse:
      type: object
      required: [reports, total, limit, offset]
      properties:
        reports:
          type: array
          items: { $ref: '#/components/schemas/ReportSummary' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
    IndividualReportListResponse:
      type: object
      required: [reports, total, limit, offset]
      properties:
        reports:
          type: array
          items: { $ref: '#/components/schemas/IndividualReportSummary' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PayloadTooLarge:
      description: Payload too large
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    OAuthDisabled:
      description: GitHub OAuth not configured
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...